Add files via upload
The "model.h5" file is likely a trained deep learning model in the Hierarchical Data Format (HDF5) format. It contains the weights, architecture, and configuration of a Convolutional Neural Network (CNN) model specifically designed for fruit classification. This file allows you to load and utilize the pre-trained model without having to train it from scratch........



Dataset: - https://www.kaggle.com/datasets/sriramr/fruits-fresh-and-rotten-for-classification

download this dataset if you want to train the model